feat: yes done fucking seeding db
