+++ database/factories/user.factory.ts
@@ -13,5 +13,6 @@ export default setSeederFactory(UserEntity, (faker) => {
   user.birthDate = faker.date.birthdate();
   user.address = `${faker.location.streetAddress()}, ${faker.location.city()}, ${faker.location.state()}, ${faker.location.country()}`;
   user.avatar = faker.datatype.boolean() ? faker.image.avatar() : null;
+  user.createdAt = faker.date.past({ years: 2 });
   return user;
 });
