+++ justfile
@@ -2,6 +2,7 @@ default: run-api-server-dev
 
 alias c := type-check
 alias db := run-docker-db
+alias dbr := reset-database
 
 restore:
   npm i
@@ -13,6 +14,10 @@ run-api-server-prod:
 [doc('run nestjs dev')]
 run-api-server-dev:
   npm run start:dev
+
+[doc('reset db')]
+reset-database:
+  npm run db:reset
 
 [doc('run docker compose local postgres')]
 run-docker-db:
